requirements.txt
opencv-python
mediapipe
numpy
googletrans==4.0.0-rc1
gTTS
pygame
Pillow
